Modify GitHub action to copy directories recursively
In our existing GitHub action, we had an issue while trying to copy directories in the 'incidents/' directory to 'public/incidents/'. The action failed when it encountered a directory because we didn't include the '-r' flag to copy directories recursively.

This commit resolves this issue by modifying the action to check if a source is a directory before deciding whether to use 'cp -r' or just 'cp'. Now, if the source is a directory, we use 'cp -r' to copy it and all its contents. If it's a file, we use 'cp' without the '-r' flag.